+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 0.1.1
 Release TBD
+- Provide the ability to add custom stopwords
 - Queue database writes during batch indexing to improve performance
 - Handle multiple child IDs in plugin_itemdeleted function
 - Do not index disabled or draft documents